refactor: 컴포넌트 계층 구조 정리
