+++ frontend/apps/schedule/components/ScheduleEdit/ScheduleEditForm/ScheduleMemo/ScheduleMemo.tsx
@@ -1,7 +1,7 @@
 import { CREATE } from '@SCH/constants/CREATE';
 import { Label, Textarea } from '@amadda/external-temporal';
 import { useContext } from 'react';
-import { ScheduleFormContext } from '../ScheduleEdit';
+import { ScheduleFormContext } from '../ScheduleEditForm';
 
 export function ScheduleMemo() {
   const { values, handleChange } = useContext(ScheduleFormContext);
