+++ frontend/apps/schedule/components/ScheduleEdit/ScheduleEditForm/ScheduleName/ScheduleName.tsx
@@ -1,7 +1,7 @@
 import { CREATE } from '@SCH/constants/CREATE';
 import { Input, Label } from '@amadda/external-temporal';
 import { useContext } from 'react';
-import { ScheduleFormContext } from '../ScheduleEdit';
+import { ScheduleFormContext } from '../ScheduleEditForm';
 
 export function ScheduleName() {
   const { values, handleChange } = useContext(ScheduleFormContext);
