+++ frontend/apps/schedule/components/ScheduleEdit/ScheduleEditFrame/ScheduleEditFrame.css.ts
@@ -1,0 +1,4 @@
+import { colors } from '@amadda/external-temporal';
+import { style } from '@vanilla-extract/css';
+
+export const BASE = style({});
