+++ frontend/apps/schedule/pages/create/index.tsx
@@ -1,5 +1,10 @@
-import { ScheduleEdit } from '@SCH/components/ScheduleEdit/ScheduleEdit';
+import { ScheduleEditForm } from '@SCH/components/ScheduleEdit/ScheduleEditForm/ScheduleEditForm';
+import { ScheduleEditFrame } from '@SCH/components/ScheduleEdit/ScheduleEditFrame/ScheduleEditFrame';
 
 export default function Page() {
-  return <ScheduleEdit />;
+  return (
+    <ScheduleEditFrame>
+      <ScheduleEditForm />
+    </ScheduleEditFrame>
+  );
 }
